Remove calculations file, fix imports
